# generated/swagger/Swagger.json
{
  "host": "dropit.deliveries.admin.co.il",
  "info": {
    "title": "DeliveriesManageAPI",
    "contact": {},
    "version": "0.01",
    "description": "Deliveries control system",
    "termsOfService": ""
  },
  "paths": {
...
    "/Users/ResolveAddress/{UserID}": {
      "post": {
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/FormattedAddressResponse"
            },
            "description": "Status 200"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "Error Response"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "UserID",
            "type": "integer",
            "required": true
          }
        ],
        "x-AuthType": [
          "None"
        ],
        "operationId": "UsersResolveAddress",
        "x-serviceID": 15,
        "x-methodName": "ResolveAddress",
        "x-serviceName": "Users"
      }
    },
...
    "/Users/FormattedAddressUpdate/{UserID}/{FormattedAddress}": {
      "post": {
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/FormattedAddressResponse"
            },
            "description": "Status 200"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "Error Response"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "FormattedAddress",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "UserID",
            "type": "integer",
            "required": true
          }
        ],
        "x-AuthType": [
          "None"
        ],
        "operationId": "UsersFormattedAddressUpdate",
        "x-serviceID": 15,
        "x-methodName": "FormattedAddressUpdate",
        "x-serviceName": "Users"
      }
    }
  },
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "basePath": "/v1",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "definitions": {
    "Error": {
      "type": "object",
      "properties": {
        "Description": {
          "type": "string",
          "example": "bad username",
          "description": "General ui of the error"
        },
        "ReasonPhrase": {
          "type": "string",
          "example": "1, 4",
          "description": "Specific reason for the error"
        }
      }
    },
...
    "FormattedAddressResponse": {
      "type": "object",
      "properties": {
        "Status": {
          "type": "integer",
          "example": 200
        },
        "Description": {
          "type": "string",
          "example": "UserID: XXX FormattedAddress Updated"
        }
      }
    }
  },
  "securityDefinitions": {}
}